Update the owner reference of resources, if the reference is not correct. (#230)
This addresses the situation when a resource, that a SolrCloud or SolrPrometheusExporter is supposed to control, has a different OwnerReference with controller= true. If this is found, then the controlling owner resource will be downgraded to controller=false, and a controlling OwnerReference will be created for either the SolrCloud or SolrPrometheusExporter.