Pass exporter args as env vars where possible (#639)
In recent versions of Solr (>=8.8) the prometheus exporter has an
env-var equivalent for each CLI argument taken by its start script.

This commit switches the operator over to providing values using this
env-var capability where possible.  Hopefully this makes the created
resources a little cleaner and easier-to-read for administrators and
users.