Clarify lack of 'standalone' Solr support
The Solr-operator only has support for running Solr in "cloud" mode, as
evidenced by the required ZK coordinates and even the name of the
'solrcloud' custom resource.  But there are few explicit mentions of
this in the README or project docs.

This commit tweaks the docs to make this slightly more explicit.