Bump 'default' Solr version to 9.9.0
This leaves a few vistigial bits of operator code that exist in support
of 8.11.  And this commit doesn't update our documentation to indicate
that 8.11 is no longer supported.  Those can be handled in a subsequent
commit or PR.

But otherwise this commit does everything necessary to make 9.9.0 the
"default" and the version assumed by our docs.

Resolves #806